+++ Modules/Microsoft365DSC/DSCResources/MSFT_SCAutoSensitivityLabelPolicy/MSFT_SCAutoSensitivityLabelPolicy.psm1
@@ -217,14 +217,14 @@ function Get-TargetResource
 
         $ExchangeSenderMemberOfExceptionValue = @()
         if (-not [System.String]::IsNullOrEmpty($policy.ExchangeSenderMemberOfException))
-        {            
+        {
             $ExchangeSenderMemberOfExceptionValue = $policy.ExchangeSenderMemberOfException.Name
         }
         $result.Add('ExchangeSenderMemberOfException', $ExchangeSenderMemberOfExceptionValue)
 
         $ExchangeSenderExceptionValue = @()
         if (-not [System.String]::IsNullOrEmpty($policy.ExchangeSenderException))
-        {            
+        {
             $ExchangeSenderExceptionValue = $policy.ExchangeSenderException.Name
         }
         $result.Add('ExchangeSenderException', $ExchangeSenderExceptionValue)
@@ -669,7 +669,7 @@ function Test-TargetResource
     )
 
     #region Telemetry
-    $ResourceName = $MyInvocation.MyCommand.ModuleName -replace 'MSFT_', ''
+    $ResourceName = $MyInvocation.MyCommand.ModuleName.Replace('MSFT_', '')
     $CommandName = $MyInvocation.MyCommand
     $data = Format-M365DSCTelemetryParameters -ResourceName $ResourceName `
         -CommandName $CommandName `
@@ -677,21 +677,9 @@ function Test-TargetResource
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
 
-    Write-Verbose -Message "Testing configuration of Sensitivity label for $Name"
-    $CurrentValues = Get-TargetResource @PSBoundParameters
-
-    $ValuesToCheck = $PSBoundParameters
-
-    Write-Verbose -Message "Current Values: $(Convert-M365DscHashtableToString -Hashtable $CurrentValues)"
-    Write-Verbose -Message "Target Values: $(Convert-M365DscHashtableToString -Hashtable $ValuesToCheck)"
-
-    $TestResult = Test-M365DSCParameterState -CurrentValues $CurrentValues `
-        -Source $($MyInvocation.MyCommand.Source) `
-        -DesiredValues $PSBoundParameters `
-        -ValuesToCheck $ValuesToCheck.Keys
-
-    Write-Verbose -Message "Test-TargetResource returned $TestResult"
-    return $TestResult
+    $result = Test-M365DSCTargetResource -DesiredValues $PSBoundParameters `
+                                         -ResourceName $($MyInvocation.MyCommand.Source).Replace('MSFT_', '')
+    return $result
 }
 
 function Export-TargetResource
@@ -802,4 +790,3 @@ function Export-TargetResource
 }
 
 Export-ModuleMember -Function *-TargetResource
-