+++ Modules/Microsoft365DSC/DSCResources/MSFT_TeamsUser/MSFT_TeamsUser.psm1
@@ -443,4 +443,3 @@ function Export-TargetResource
 }
 
 Export-ModuleMember -Function *-TargetResource
-